+ Cursor Box and Cursor
The cursor box will only appear when you are going to place a block.
When you can't place a block at that spot, it will be red. Right now
there are no blocks to place so the Cursor Box is just there to review.

Also, when you mouse-over a button, the cursor will change